% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{displace}
\alias{displace}
\title{Add spatial displacement per frame for each track object}
\usage{
displace(population, strata, x_var = "Location_Center_X",
  y_var = "Location_Center_Y", t_var = "Metadata_timePoint")
}
\arguments{
\item{population, }{data frame storing single cell data}

\item{strata, }{column name storing the track label}

\item{x_var}{variable name / columne name used for x-coordinates}

\item{y_var}{variable name / columne name used for y-coordinates}

\item{t_var}{variable name / columne name used for time coordinates}
}
\value{
displacement
}
\description{
Add spatial displacement per frame for each track object
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tracks <- migrationminer::displace(data,"TrackObjects_Label")
}
