% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{forward_migration_index}
\alias{forward_migration_index}
\title{Compute the forward migration index of a tracked object}
\usage{
forward_migration_index(tracks, x_var = "Location_Center_X",
  y_var = "Location_Center_Y")
}
\arguments{
\item{tracks}{data frame with track objects}

\item{x_var}{variable name / columne name used for x-coordinates}

\item{y_var}{variable name / columne name used for y-coordinates}
}
\value{
forward migration index
}
\description{
Compute the forward migration index of a tracked object
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tracks <- migrationminer::displace(data,"TrackObjects_Label")
 forward_migration_index <- migrationminer::forward_migration_index(tracks)

}
