% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{mean_position}
\alias{mean_position}
\title{Get mean position}
\usage{
mean_position(tracks, strata)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{strata}{column name of track index column}
}
\value{
data frame with mean x and y positions
}
\description{
Get mean position
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 strata <- 'TrackObjects_Label'
 data <- dplyr::group_by_(data,strata)
 position <- neutrominer::mean_position(data,strata)
}
