% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{plot_windrose}
\alias{plot_windrose}
\title{Windrose style plot}
\usage{
plot_windrose(tracks, spdres = 1, dirres = 30, spdmin = 0,
  spdmax = 5, palette = "YlGnBu",
  scale_name = "speed in pixel/frame",
  title_name = "Distribution of angle and speed")
}
\arguments{
\item{tracks}{data frame with track objects}

\item{spdres}{speed resolution used for coloring}

\item{dirres}{angle resolution of the plot}

\item{spdmin}{minimum speed (used for color mapping of speed)}

\item{spdmax}{max speed (used for color mapping of speed )}

\item{palette}{color map (used for color mapping of speed)}

\item{scale_name}{legend name / title for the scale}

\item{title_name}{title name of the plot}
}
\value{
windrose_plot
}
\description{
Windrose style plot
}
\examples{
 df <- dplyr::tibble(
   Track_Speed = runif(1000, max = 5),
   Track_Angle =  runif(1000, min = -pi, max = pi)
 )
 plot_windrose(df)
}
