% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{summarize_sectors}
\alias{summarize_sectors}
\title{Summarize sector analysis per experiment}
\usage{
summarize_sectors(tracks, strata)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{strata}{column name of track index column}
}
\value{
sector_summary
}
\description{
Summarize sector analysis per experiment
}
\examples{
 data <- tibble::tibble(
   Metadata_timePoint = c(rep(0, 5),rep(1,5)),
   Location_Center_X = c(0, 0, 0, 0, 0, -1,  1,  0,  0, 0),
   Location_Center_Y = c(0, 0, 0, 0, 0,  0,  0,  1, -1, 1),
   TrackObjects_Label = c(1:5,1:4,1),
   Metadata_experiment = c(rep(0,10))
 )

 strata <- c('TrackObjects_Label','Metadata_experiment')
 tracks <- track(data, strata = strata)
 sector <- summarize_sectors(tracks,'Metadata_experiment')
}
