% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Compute track statistics}
\usage{
track(population, strata, x_var = "Location_Center_X",
  y_var = "Location_Center_Y", t_var = "Metadata_timePoint")
}
\arguments{
\item{population, }{single cell data}

\item{strata, }{column name storing the track label}

\item{x_var}{variable name / columne name used for x-coordinates}

\item{y_var}{variable name / columne name used for y-coordinates}

\item{t_var}{variable name / columne name used for time coordinates}
}
\value{
track
}
\description{
Compute track statistics
}
\examples{
 data <- tibble::tibble(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 data <- dplyr::group_by_(data,"TrackObjects_Label")
 tracks <- migrationminer::track(data,"TrackObjects_Label")
}
